Moved from raw pointer to smart pointer
Implementation of the handling of the instream with a smart pointer
instead of a raw pointer. As std::cin should not be deallocated on
program termination, this currently involves a pretty ugly bit of code
in the destructor, sporting a new shared pointer on std::cin being
allocated in the heap. Yup, that ugly...

Signed-off-by: Max Franke <max@mumintroll.org>